Render the Cart
Once a `Book` record is in a visitor's `Order`, the corresponding
`LineItem` record is rendered into the page's cart.

To render the cart, and to handle the expanding and collapsing of the
cart, render a [`<details>` element][mdn-details] in the page's [`<nav>`
element][mdn-nav] containing a [`<summary>` element][mdn-summary]
element to toggle it open.

In order to apply Tailwind classes in the different [`[open]` attribute
states][mdn-open], declare an ["open" variant][tailwind-variant] to make
`open:`-prefixed utilities available. Since the number of selectors is
limited to `display:` and `position:` rules, forego writing a custom
variant, and instead declare the CSS rules directly.

To simplify the interactions with the cart in the System Test suite,
introduce the `expand_cart` test helper, which invokes
[`Capybara::Node::Finders#find`][find] with the `text:` option
containing the translated text we embedded into the cart.

[mdn-details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[mdn-summary]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/summary
[mdn-nav]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/nav
[mdn-open]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#attr-open
[tailwind-variant]: https://tailwindcss.com/docs/plugins#adding-variants
[find]: https://www.rubydoc.info/github/jnicklas/capybara/master/Capybara/Node/Finders#find-instance_method