Render the Cart
Once a `Book` record is in a visitor's `Order`, the corresponding
`LineItem` record is rendered into the page's cart.

To render the cart, and to handle the expanding and collapsing of the
cart, render a [`<details>`][details] element in the page's `<nav>`,
containing an SVG icon in the child [`<summary>`][summary] element.

To simplify the interactions with the cart in the System Test suite,
introduce the `expand_cart` test helper, which invokes [`find`][find]
with the `text:` option containing the translated `<title>` text we
embed into the cart icon's `inline_svg_tag`.

[details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[summary]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/summary
[find]: https://www.rubydoc.info/github/jnicklas/capybara/master/Capybara/Node/Finders#find-instance_method