Add a Book to the Order
This commit introduces the `Order` model to represent the items that a
visitor has added to their cart.

An `Order` record is identifiable by its [signed GlobalID][globalid],
which the `ApplicationController` uses to retrieve the correct record
via [`ActiveRecord::SignedId.find_signed`][find_signed].

When a visitor initially navigates around the site, they won't yet have
an `Order` record. Once they decide to add a `Book` record to their
`Order`, that action will create a `LineItem` record.

Once the `LineItem` record is persisted, the newly created `Order`
record's [`signed_id` value][signed_id] will be persisted to their
cookie. Subsequent requests will include that cookie, which will serve
to identify and re-use the browsing session's `Order` record.

The "Add to Cart" button is powered by a [`form_with`][form_with] that
submits `POST /books/:book_id/line_items` requests, which are routed to
the `line_items#create` controller action.

Those requests are made with an `increment` key in their [HTTP Form URL
encoded data][x-www-form-urlencoded]. The `increment` attribute
represents the amount to modify the `LineItem#quantity` by. The default
action of adding a single `Book` to an order will declare an `increment`
of `1`.

Forms constructed by `form_with` calls [will default to encoding
`data-remote="true"` attributes][remote-forms] as an integration with
[RailsUJS][]. These `<form>` submissions will be made over
asynchronously with `XMLHTTPRequest`-powered requests.

When the `LineItemsController` responds with a
[`redirect_back`][redirect_back], [Turbolinks will intervene and handle
the redirection on the client-side][turbolinks-redirect] in a way that
re-uses the existing page's `<head>` element's state.

While the Turbolinks loading might is a client-side optimization, it's
worthwhile to consider the redirect conceptually as we would a full page
reload.

Testing
---

The `LineItemsController` tests exercise the controller by making `POST`
HTTP Requests. The test cases make assertions about whether or not a new
`Order` record is created, or whether an existing `Order` record is
fetched and used.

To do so, they declare the `assert_created_order` and
`assert_no_created_order` assertions, which extend the
[assert_changes][] and [assert_no_changes][] assertions, respectively.

[globalid]: https://github.com/rails/globalid/tree/v0.4.2#signed-global-ids
[find_signed]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/SignedId/ClassMethods.html#method-i-find_signed
[signed_id]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/SignedId.html#method-i-signed_id
[form_with]: https://edgeapi.rubyonrails.org/classes/ActionView/Helpers/FormHelper.html#method-i-form_with
[x-www-form-urlencoded]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods/POST#Example
[remote-forms]: https://edgeguides.rubyonrails.org/form_helpers.html#how-do-forms-with-patch-put-or-delete-methods-work-questionmark
[RailsUJS]: https://edgeguides.rubyonrails.org/working_with_javascript_in_rails.html#remote-elements
[redirect_back]: https://edgeapi.rubyonrails.org/classes/ActionController/Redirecting.html#method-i-redirect_back
[turbolinks-redirect]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#redirecting-after-a-form-submission
[assert_changes]: https://edgeapi.rubyonrails.org/classes/ActiveSupport/Testing/Assertions.html#method-i-assert_changes
[assert_no_changes]: https://edgeapi.rubyonrails.org/classes/ActiveSupport/Testing/Assertions.html#method-i-assert_no_changes