Add tests for viewing Books in the Cart
In order to pre-populate the [browser's HTTP Cookies][mdn-cookies] with
an `Order` record's `token` value, this commit makes use of the
[`rack-back_door` middleware][rack-back_door].

Inspired by thoughtbot's [`clearance` gem][clearance], the
`rack-back_door` provides a mechanism for intervening in a System Test
request's life-cycle.

To quote the `clearance` documentation:

> Clearance includes middleware that avoids wasting time spent visiting,
> loading, and submitting the sign in form. It instead signs in the
> designated user directly. The speed increase can be
> [substantial][backdoor].

This commit re-purposes that approach to load an existing `Order` record
(referenced by the `cookie[:order_token]`) record instead of a `User`
record.

[mdn-cookies]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies
[rack-back_door]: https://github.com/seanpdoyle/rack-back_door/tree/v0.1.2
[clearance]: https://github.com/thoughtbot/clearance/tree/v2.1.0#fast-feature-specs
[backdoor]: http://robots.thoughtbot.com/post/37907699673/faster-tests-sign-in-through-the-back-door