Progressively enhance persistent elements
Permanence:
===

In the interest of improving the end-user experience, this commit
extends the page's Cart and Flash Messages to persist across
[`turbolinks:load`-triggered page reloads][turbolinks:load].

To integrate with [Turbolinks support for Persistent
Elements][permanent], elements must both:

* declare the `data-turbolinks-permanent` attribute
* declare a unique [`id`][id] attribute

According to the Turbolinks documentation:

> Before each render, Turbolinks matches all permanent elements by `id`
> and transfers them from the original page to the new page, preserving
> their data and event listeners.

This style of long-lived persistence and state dovetails nicely with
elements decorated by [StimulusJS controllers][stimulusjs].

For example, since the cart is represented by the [`<details
id="order">` element][details], there isn't any user-land JavaScript
required to manage its display state. <detail> elements marks themselves
with the [`[open]`][open] attribute when expanded, and remove the
`[open]` attribute when collapsed.

This integrates with Turbolinks' notion of Permanent Elements, since its
internal state is represented in the DOM, so transposing its Node across
[`document.body` trees][document.body] inherently embeds its state.

This commit marks the `<details id="order">` element with
`data-turbolinks-permanent` so that its state is persisted across page
loads.

To imagine this in action, consider adding a Book to the Cart with the
list of currently ordered Books expanded. Clicking on an "Add to Cart"
button elsewhere in the page and having the Cart close itself could be
jarring. By persisting the expanded state of the Cart during
interaction, the visitor's continuous shopping is uninterrupted.

However, there are caveats.

Transforming the Cart into a long-lived element puts it at risk of
drifting out of synchronization with the rest of the page. To return to
the previous example, a visitor might be surprised to see the Cart close
itself when adding a Book, but they'd likely be even _more_ frustrated
if the Book they've just added to their Cart never materialized in their
list of `LineItem` records.

To account for that synchronization, this commit introduces a Stimulus
controller that is attached to elements bearing the
`[data-controller="permanent"]` attribute along with the
`[data-action="turbolinks:before-render@document->permanent#cloneChildren"]`,
which [routes `turbolinks:before-render` events][action-routing] fired [on the
`document`][@document] to the `cloneChildren` action declared within the
`"permanent"` controller.

When marked "permanent", an element will persist its own state (the
presence or absence of the `open` attribute, in the Cart example), but
_will_ clone over its new children (plucked out of its corresponding
node in the `newBody`, identified by its unique `id` attribute)
available in the `turbolinks:before-render` event's data.

This approach benefits from both client-side persistence of element
state, while still receiving newly server-generated state.

A similar pattern is applied to the page's Flash Messaged, represented
by the `<div id="flashes">` element.

When marked with `data-permanent-append="true"`, permanent elements like
`<div id="flashes">` will [_append_ the new children after existing
children][insertAdjacentHTML], instead of replacing their existing ones.

When taking actions across a long-lived session (like adding then
removing a book in rapid succession), sequential Flash Messages could be
rendered adjacent to one another, to more clearly communicate the
impacts of the visitor's actions.

The `[data-permanent-append]` attribute is de-serialized from the
element's attributes as a [`Boolean`][boolean] thanks to the new Values
API made available in [Stimulus 2.0][stimulus-values].

Transience
===

The [typical Stimulus controller life-cycle][lifecycle] involves a call
to `connect()` when the corresponding element is detected in the DOM,
and a `disconnect()` call when the corresponding element is removed from
the DOM.

These book-end hooks serve to configure the element during `connect()`,
and to de-construct and cleanup after the element during `disconnect()`.

In the case of the `<div id="flashes">` element, permanence is at odds
with the nature of the element's use case. The intention behind showing
a Flash message is to quickly (and temporarily) communicate a message to
the user that explains the impacts of their most recent actions.

Persisting a Flash Message on a page for too long could become
distracting, or (at worst) disorienting.

This commit also introduces a controller for elements annotated with the
`[data-controller="temporary"]` attribute.

When marked as `"temporary"`, an element schedules itself to be removed
by invoking [`remove()`][remove] from within a
[`setTimeout()`][setTimeout], configured to fire in the future.

By default, the element will remove itself in 5 seconds. That value can
be configured by declaring a millisecond timeout value in the controller
element's `[data-temporary-delay="..."]` attribute.

[turbolinks:load]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#full-list-of-events
[permanent]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#persisting-elements-across-page-loads
[stimulusjs]: https://stimulusjs.org/handbook/introduction
[details]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
[open]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details#attr-open
[document.body]: https://developer.mozilla.org/en-US/docs/Web/API/Document/body
[action-routing]: https://stimulusjs.org/reference/actions
[@document]: https://stimulusjs.org/reference/actions#global-events
[boolean]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Boolean
[lifecycle]: https://stimulusjs.org/handbook/managing-state#lifecycle-callbacks-explained
[stimulus-values]: https://stimulus.hotwire.dev/reference/values
[insertAdjacentHTML]: https://developer.mozilla.org/en-US/docs/Web/API/Element/insertAdjacentHTML
[id]: https://developer.mozilla.org/en-US/docs/Web/HTML/Global_attributes/id
[remove]: https://developer.mozilla.org/en-US/docs/Web/API/ChildNode/remove
[setTimeout]: https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/setTimeout