Persist Scroll Position when Ordering Books
Declare `data-scroll-permanent` to persist scroll position across
`<form>` submissions.

The Stimulus controller backing elements annotated with
`[data-controller="scroll"]` (in this case, the document's `<body>`
element) is inspired by ["How To: Turbolinks 5 Scroll Position
Persistence"][scroll-position].

By definition, the `<body>` element is the ancestor to all
browser-rendered and interactive elements in the DOM.

By declaring `[data-action]` attributes routing [`submit`][form-submit]
and [`turbolinks:load`][turbolinks:load] events to the
`scroll#trackPosition` and `scroll#restorePosition`, respectively,
users' current scroll positions are persisted to the body as the
`[data-scroll-top]` attribute.

Turbolinks page loads don't replace the `<body>` element itself, only
its children. In effect, this makes the `<body>` a Persisted Element,
like the Cart and Flash Messages, and benefits in the same ways.

It is a deliberate design decision to _only_ save the scroll position
during Turbolinks page loads initiated by `<form>` elements annotated
with the `[data-scroll-permanent="true"]` attribute. Making this
behavior opt-in as opposed to opt-out means that the default behavior is
a battle-tested full page navigation. To that end, once a scroll
position is restored, the attribute tracking that value is unset, and
default navigation behavior is restored.

[scroll-position]: https://medium.com/@kosovacsedad/how-to-turbolinks-5-scroll-position-persistence-6e4435a60b2e
[form-submit]: https://developer.mozilla.org/en-US/docs/Web/API/HTMLFormElement/submit_event
[turbolinks:load]: https://github.com/turbolinks/turbolinks/tree/v5.2.0#full-list-of-events