Remove a Book from the Cart
Add a [`<form data-remote="true">` element][data-remote] to each
rendering for an `Order` record's related `LineItem` records.

The `<form>` submits to the new `DELETE /line_items/:id` route, which
removes the `LineItem` record that associates the `Book` record to that
visitor's `Order` record.

Testing
---

At this point, several of our System Tests exercise a similar process:
visiting a `Book` and adding it to the `Order`. This commit extracts
some common helpers with communicative names to describe that process.
Those helpers are implemented with the new `#label` and `#submit`
helpers, which abstract over the `ActionView::Helpers::FormBuilder`
label and submission button translation interface.

[data-remote]: https://edgeguides.rubyonrails.org/working_with_javascript_in_rails.html#form-with