Render available Books
As the next step in the project's setup, introduce the `Book` model,
backed by the `books` database table and pre-seeded by the
[test/fixtures/books.yml][] ActiveRecord [fixtures file][FixtureSet].

To ensures that the fixture data is available in the `development`
environment, call the `db:fixtures:load` Rails task as part of the
`bin/setup` script.

A `Book` record has:

* a `String` title
* an `Integer` price in cents USD
* a `DateTime` value for both the time it was created at and last
  updated at

To ensure the dynamic behavior is functioning correctly, this diff
includes a [System Test][] to drive the behavior.

Styling the list
---

With the newly available Tailwind utility classes, style the page by
chaining utility classes across various elements.

To ensure that our page's responsive styles apply on differently sized
devices, declare a [Viewport `<meta>` element][viewport-meta] to hint
to the browser:

> A typical mobile-optimized site contains something like the following:
>
> `<meta name="viewport" content="width=device-width, initial-scale=1">`
>
> The `width` property controls the size of the viewport. It can be set
> to a specific number of pixels like `width=600` or to the special
> value `device-width`, which is the width of the screen in CSS pixels
> at a scale of 100%. (There are corresponding `height` and
> `device-height` values, which may be useful for pages with elements
> that change size or position based on the viewport height.)
>
> The `initial-scale` property controls the zoom level when the page is
> first loaded. The `maximum-scale`, `minimum-scale`, and `user-scalable`
> properties control how users are allowed to zoom the page in or out.
>
> Usage of `user-scalable=no` can cause accessibility issues to users
> with visual impairments such as low vision.

[System Test]: https://edgeguides.rubyonrails.org/testing.html#system-testing
[viewport-meta]: https://developer.mozilla.org/en-US/docs/Mozilla/Mobile/Viewport_meta_tag
[app/views/layouts/application.html.erb]: #app/views/layouts/application.html.erb
[test/fixtures/books.yml]: #test/fixtures/books.yml
[FixtureSet]: https://edgeapi.rubyonrails.org/classes/ActiveRecord/FixtureSet.html